Restyle about and index
I made a bunch of changes to the css and the html to make it more clear. I changed the spacing in the index html to be line with the standard spacing for elements. I also added in the arrows and css styling for those. I made it compatible with the about page and I'm currently working on integrating the about css into the standard style css. I also made some minor changes to the about page to make it more in line with the style css declaration blocks