Style pages
I styled all the pages to match the design better. I made a few design tweaks along the way but the only thing that is majorly different is the size of the components no longer scale with the height. I also added in a home button and changed the behavior of the navigation links