Refactor timeout
I refactored the way that the timeout function worked. Now it can cancel without having a nested listener. That implementation was weird. There is still a bug where sometimes the text won't popup. I'm not sure why this is happening. I think it has something to do with moving between dates too quickly. Probably won't be a probably if an actual user. But you never know. Maybe a user can be looking for a specific date and they're going through the dates quickly and the date just doesn't show... hmmm. probably is a problem then...